common: move all *.h, except utils.h and bi_defines.h back to battle_interface/src
This partially reverts commit 7169f0049c279f1d947f587971058184b6e1eee5
