+++ src/libs/battle_interface/src/battle_sign.h
