+++ src/libs/battle_interface/src/land/battle_man_command.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "battle_command.h"
+#include "../battle_command.h"
 
 class BIManCommandList : public BICommandList
 {
