+++ src/libs/battle_interface/src/sea/i_battle.cpp
@@ -13,8 +13,8 @@
 #include "../land/i_boarding_status.h"
 #include "../land/battle_land.h"
 
-#include "active_perk_shower.h"
-#include "log_and_action.h"
+#include "../active_perk_shower.h"
+#include "../log_and_action.h"
 #include "../item_entity/item_entity.h"
 #include "../world_map_interface/interface.h"
 #include "../spyglass/spyglass.h"
